[NPUW] Split test-engine model builder into per-component files
Pure file-split refactor of the NPUW test-engine model builder. No
behavior changes, no public API changes.

Splits the ~2.9k-line model_builder.{hpp,cpp} into per-component
sibling files so future mask / attention / FFN work touches only the
relevant file.

New files under tests/functional/behavior/npuw/test_engine/models/:
  model_builder_types.hpp     — shared typedefs + KVCache structs.
  model_builder_internal.hpp  — internal seeded-PRNG helpers and
                                constants.
  model_builder_weights.{hpp,cpp}  — FloatWeight, CompressedWeight,
                                     DCOff sugar.
  model_builder_norm.{hpp,cpp}     — LayerNorm, RMSNorm.
  model_builder_rope.{hpp,cpp}     — RoPE structs, position_ids
                                     helpers, learned positional
                                     embedding.
  model_builder_ffn.{hpp,cpp}      — SwiGLU, GELU, MoEFFN.
  model_builder_attention.{hpp,cpp} — Attention struct + helpers.
  model_builder_masks.{hpp,cpp}    — padding / causal / whisper causal
                                     masks + cache_position.

model_builder.{hpp,cpp} keeps the ModelBuilder class, top-level
build_llm / build_whisper_* / build_embedding_encoder, and shared
utility helpers (make_linear, make_embedding, make_lm_head,
make_conv1d, make_transformer_layers, make_pre_norm_layer,
make_cross_attn_decoder_layer). The public hpp re-exports the split
headers so existing includes keep working.

tests/unit/CMakeLists.txt updated to add the new .cpp files.
Functional tests pick the new sources up via the existing GLOB in
tests/functional/CMakeLists.txt.
