+++ src/plugins/intel_npu/tests/functional/behavior/npuw/test_engine/models/model_builder_norm.hpp
@@ -1,0 +1,44 @@
+// Copyright (C) 2018-2026 Intel Corporation
+// SPDX-License-Identifier: Apache-2.0
+//
+
+#pragma once
+
+#include <string>
+
+#include "openvino/core/node.hpp"
+#include "openvino/core/type/element_type.hpp"
+
+namespace ov {
+namespace test {
+namespace npuw {
+
+struct LayerNorm {
+    size_t hidden_size;
+    ov::element::Type precision;
+    float eps;
+
+    LayerNorm(size_t hs, ov::element::Type prec = ov::element::f32, float e = 1e-5f)
+        : hidden_size(hs),
+          precision(prec),
+          eps(e) {}
+
+    ov::Output<ov::Node> operator()(const ov::Output<ov::Node>& input, const std::string& name) const;
+};
+
+struct RMSNorm {
+    size_t hidden_size;
+    ov::element::Type precision;
+    float eps;
+
+    RMSNorm(size_t hs, ov::element::Type prec = ov::element::f32, float e = 1e-5f)
+        : hidden_size(hs),
+          precision(prec),
+          eps(e) {}
+
+    ov::Output<ov::Node> operator()(const ov::Output<ov::Node>& input, const std::string& name) const;
+};
+
+}  // namespace npuw
+}  // namespace test
+}  // namespace ov
